license: replace violating MPL 1.1 with GPL-3.0-or-later
Nominally, only the sensors part of the API was explicitly under the
GPL-3.0-or-later. The original license on the other parts was the
MPL-1.1 according to the license information in the file headers.

It has since turned out that even the MPL-1.1 parts include
GPL-2.0-or-later code, which is a license violation.

Mixing MPL-1.1 and GPL-3.0 code is a license violation in itself, but
because the MPL-1.1-licensed code is itself a derivative work of the
GPL-2.0-or-later code by the LEGO Group, as a whole it should have been
distributed under the GPL 2.0-or-later. The GPL-2.0-or-later is
compatible with the GPL-3.0-or-later, so the project is actually
only redistributable under the GPL-3.0-or-later. Change license
information to indicate so. Also add a note to the README that new
contributions over the pull request mechanism should be
GPL-2.0-or-later, so we may in future migrate to GPL-2.0-or-later if all
copyright holders agree.

More information at #34.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>