lcd: add function for displaying splash screens (#27)
The new LcdSplash() function will forward its arguments
to the dLcdDrawPicture() function. This function
is optimized for 8-pixel aligned pictures, so rendering
will be fast. The arguments for this function can be obtained
from a XBM file.

This function can then be used for displaying loading-time
splash screens. For that to work, it is necessary to register
a new constructor function that will run before the builtin
library initializer. In there, it is necessary to initialize
LCD module manually and then the splash can be rendered.

Example:

#include <ev3.h>
#include "loading.xbm"

void __attribute__((constructor(101))) MyFancyInit (void)
{
    LcdInit();
    LcdSplash(1, 0, 0, loading_width, loading_height, loading_bits);
}

int main() {
    Wait(2000);
}