feature(lcd): add functions for manipulating text cursor
This commit adds functions for setting and getting X and Y
coordinates of the text cursor. This partially duplicates
the functionality available in *Printf() functions; however
the added convenience is worth it. Also, the code in *Printf()
functions accesses a cached local variable instead of
the global one, so the code there would have to be modified
not to cache the variable.

Relates to #39