implement individual channels reading for HT IR V2
Add function to read the 5 individual channels of the HT Ir Seeker V2 sensor.
This new function has NOT been tested on a real device.

Since for the HT IR V2 we don't need to keep track of the current sensor mode,
the function ReadHTIrV2Sensor has been modified to not store the current mode